delcore30m: Add DSP code profiling
If profile bit is set in flag of job struct, it is posssible to
performance measurement for this job. DSP can write 32bit label
and 32bit value of TOTAL_CLK_CNTR in buffer, whose start address
in A5. Label and TOTAL_CLK_CNTR must be written before and after
code section for which performance measurement is necessary.

Example of profiling results are given below:

  Tag 1: calls = 238, min/max/avg = 150/12409/5603
  Tag 2: calls = 476, min/max/avg = 737/812/775
  Tag abab: calls = 119, min/max/avg = 10524/12446/12406

Issue: #MCOM02SW-1111
Change-Id: I2fc2ca34c1d0ede35db5a8679c1d4ed87606f772