[media] avico: Move clearing of RUN and DONE
RUN, DONE, IMRDY registers have got zero values when VPU is reset
so they shouldn't be cleared in start_streaming().

EVENTS register depends on DONE registers. For example, EV0/EV1
couldn't be set to 0 if DONE register of channel 0/1 is 1. So we
should clear DONE registers before clearing of EVENTS.

When VPU is stopped, EVENTS register should be cleared to deassert
interrupts. DONE registers must be cleared before EVENTS. RUN
registers must be cleared before DONE to stop DMA otherwise it will
start working because all conditions are met for that.

We are going to use channels 0 and 1 to save and restore hardware
thread context so EV0 and EV1 are cleared.

Issue: #MCOM02SW-750
Issue: #MCOM02SW-1300
Issue: rf#11257
Change-Id: I318ebbd1df43dafe037ca0d7fa057433da2e1475