[media] avico: Finish job if out_size is zero
Job won't be finished if out_size is zero for last row of MBs because
avico_dma_out_callback() won't be called. out_size could be zero
when AECUR is wrapped and becomes equal to A0E. This happens when
encoded data written to bounce buffer is greater or equal to
BOUNCE_BUF_SIZE.

This patch marks capture buffer erroneous and finishes job when
out_size is zero for last row of MBs.

Issue: #MCOM02SW-750
Issue: #MCOM02SW-1086
Change-Id: Idba3d9b7668d861cb74b050ce0000a0a9d2e1eb5