[media] avico: Save and restore mb_cur and mb_ref structures
The encoder loads the first part of a reference frame for the next
input frame from external memory to mb_ref structures in VRAM at
the end of current job. The first part includes:

1. Bottom 16x8 partitions of MBs of row (NY - 1) and top 16x8
   partitions of MB0 of row 0. NY - number of rows of MBs in an
   input frame.
2. Bottom 16x8 partition of MB0 of row 0 and top 16x8 partition of
   MB0 of row 1.

Also, the encoder stores the final part of a reconstructed frame from
mb_cur structures in VRAM to external memory at the start of the next
job. The final part includes MBs of row (NY - 1).

These cause artefacts in video when two or more processes use the
same encoder because encoder uses first and final parts of a
reference frame from other context.

This patch:

* allocates buffers per context in DDR to store mb_cur and mb_ref
  structures which contain first and final parts of a reference
  frame;
* saves mb_cur and mb_ref structures to buffers at the end of a job
  using channels 0 and 1 of VDMA;
* restores mb_cur and mb_ref structures at the start of a job using
  channels 0 and 1 of VDMA.

Issue: #MCOM02SW-750
Change-Id: I69ea2721cef9deb477d79e8699e87b6dc6d2d0ba