net: arasan: Change TX mechanism
TX mechanism consists of an arasan_gemac_start_xmit()
and a part of IRQ handler. So far, the arasan_gemac_start_xmit()
does following steps:

* Releases all SKBs that have been transferred.
* Creates DMA descriptor for SKB.
* Inserts DMA descriptor to own queue.
* Kicks the DMA.

Note that a socket has own queue that is independent from
a driver queue. So, if the socket queue becomes full and the
driver queue is not full, then the driver hits to a loop,
because it cannot flush the socket queue until a next SKB
arrives. The next SKB from this queue will not arrive
because this queue is frozen. Another sender can break this
loop (for example the ARP subsystem), but the socket
performance is decreased. This scenario can de avoided if
the driver queue will be less than the socket queue,
but it seems wrong. For this reason performance degradation
has been observed if driver queue was increased.

Now, a descriptor unavailable IRQ is used. It guarantees that
TX ring will be freed when no more TX descriptors are available
and socket queue will be defrozen.

Issue: #MCOM02SW-1214
Change-Id: Ic507cfef23b7a8e9187f3010a3a125c204efbd4d