dt-bindings: gpio: Rename 'skip-gpios' to avoid dtc warning
dtc produces the next warning when compiling DTS file that contains
'skip-gpios' property:

  Warning (gpios_property): /gpio@38034000/gpio-controller@3: Missing property '#gpio-cells' in node uart0_pclk or bad phandle (referred from skip-gpios[0])

This is because any property which includes 'gpios' in a name
interpretes by the compiler as a phandle to GPIO node. To avoid
this warning one renames 'skip-gpios' property to 'skip-gpio-list'.
The commit renames the property in the documentation.

Issue: #MCOM02SW-1244
Change-Id: Ida5dca5a79dad3eb077bbbde745347c7b639da37