[media] avico: Fix async work of SDMA and VDMA
Currently, availability of bounce buffer (whether SDMA finished
copying data) is not checked when we continue processing of the frame
(start of VDMA channels 2 and 3). This leads to data overwrite in
bounce buffers and artefacts in encoded video.

Now, VDMA channels 2 and 3 will be run in avico_irq() if there are
available bounce buffers otherwise VDMA channels 2 and 3 will be run
in avico_sdma_callback(). avico_sdma_callback() is called when
reconstructed data was copied. A transaction for encoded data is
submitted before transaction for reconstructed data so encoded data
is copied before reconstructed data and it is enough to use only
avico_sdma_callback() in order to wait for available bounce buffers.

Issue: #MCOM02SW-1125
Change-Id: I7003a5fcd10d67fb7dca383e6e343ae66868b9a5