[media] avico: Add support for chroma_qp_index_offset setting
The chroma_qp_index_offset parameter of PPS specifies the offset
that shall be added to QPY for addressing the table of QPC values
for the Cb chroma component.

The patch adds support for the control which allows to set
chroma_qp_index_offset parameter. When a value of the parameter is
changed a PPS with the value is written.

Issue: #MCOM02SW-1327
Issue: rf#11401
Change-Id: I2397b5cf55e598c553371a7f084389b63c5ff26a