[media] avico: Fix double free
dma_free_coherent() could be called twice for the same pointer. The
first call would be in the start_streaming(), the second call would
be in the stop_streaming(). This can lead to undefined behaviour
and exceptions.

A stop_streaming() will be called for a V4L2-interface only if
start_streaming() for the same V4L2-interface returns 0.

The avico_start_streaming() returns nonzero value only for one of
the V4L2-interfaces, so the avico_stop_streaming() would be called
for the other interface.

In order to prevent the double free, capon/outon must be 0 for
capture/output interface when start_streaming() is going to return
nonzero value.

Issue: #MCOM02SW-750
Change-Id: I04099642b3269dd93e030597fd4d5a87ae78bd54