[media] avico: Support height and width multiple of 2
The driver supports only input frames with height and width multiple
of 16 so input video with resolution 1920x1080 isn't supported.

The patch forces bytesperline and sizeimage (struct v4l2_pix_format)
and adds cropping parametrs to SPS in order to allow height and width
to be multiple of 2.

Issue: #MCOM02SW-1450
Change-Id: Id5e9193c797b71aaa119b29886679060485e988d