mmc: sdhci-mcom02: Support Tap Delay from Device Tree
Tap delay for sdcard_clk hardcoded in U-Boot/Linux code.

Add support to read tap delay values from DT.

For each mode, in DT added property with tap delay pair. If for specific
mode property in DT exists, it will be read and set for sdcard_clk.
If value in tap delay pair is set to <(-1)>, it will be disabled.
If property for specific mode does not exists in DT, tap delay will be
disabled for that mode.

Supported properties:

* mcom02,clk-phase-mmc-hs
* mcom02,clk-phase-sd-hs
* mcom02,clk-phase-uhs-sdr12
* mcom02,clk-phase-uhs-sdr25
* mcom02,clk-phase-uhs-sdr50
* mcom02,clk-phase-uhs-sdr104
* mcom02,clk-phase-uhs-ddr50
* mcom02,clk-phase-mmc-ddr52
* mcom02,clk-phase-mmc-hs200

This commit is based on discussion from Linux Kernel Mailing List [1].

[1] https://lore.kernel.org/patchwork/patch/1086927/

Issue: #MCOM02SW-1318
Change-Id: I4b773c3f0060e2658ddf91f82b36fd86b979e799