elvees-swic: Reset transmission speed if connection is lost
According to SpaceWire standard ECSS-E-ST-50-12C when connection
is lost, link TX bit rate must be forced to 2-20 Mbit/s.

This commit ensures that when connection is lost TX speed is set
to 4.8 Mbit/s. Once connection is recovered TX speed is set
to the previous value.

Issue: #MCOM02SW-1238
Change-Id: If2a027997df268bc92df4e1387291737aac28a08