[media] avico: Allocate decoder buffers in different DDRs
If reference/decoded and encoded buffers are located in the same DDR,
VPU may hang.

To implement workaround the patch puts buffers for decoder in
different DDRs properly. An information about memory reserved
in different DDRs is read from device-tree and the memory pools
are created. Two buffers for reference/decoded frame and two buffers
of encoded data are allocated in different DDRs. An encoded data is
copied to the proper DDR and the VRAM is initialized properly with
reference data before decoding is started.

This workaround requires that two DDRs are accessible on device,
so currently only Salute-EL24PM1/Salute-EL24PM2 supports decoding.

Issue: rf#1382
Issue: #MCOM02SW-1573
Change-Id: I6f8e6cfd83f0a1a318f0d4d6273d98c1f67cdd90