[media] avico: Rename freebits to bitsleft
It is useful to keep number of remaining bits in current byte when
video is decoded, so member "freebits" in bitstream structure is
renamed to more common "bitsleft". When it is encoding "bitsleft"
will mean number of empty bits in current byte. When it is decoding
"bitsleft" will mean number of remaining bits in current byte.

Change-Id: I3837ecea7b2f8de4288b350069bf2a541948166a