fbdev: vpoutfb: Implement fb_destroy() properly
VPOUT should be switched off in fb_destroy() (before resources are
freed) and in fb_set_par() (before VPOUT is reconfigured). VPOUT
isn't stopped (lcdcsr.RUN=0) until end of frame (VSYNC1->0) so
SYNC_DONE interrupt is used to wait for end of frame. There is 50 ms
timeout for SYNC_DONE interrupt waiting (the same timeout is in
vpout-drm).

Overall timeout for VPOUT stopping and clearing is increased from
40 ms to 90 ms. This fixes "Failed to clear FIFO" error.

A timeout isn't changed, when VPOUT is switched on, because VPOUT
doesn't have to be stopped before clearing.

reset_tasklet is disabled in fb_destroy() because the tasklet
shouldn't be executed while VPOUT is switched off.

Issue: #MCOM02SW-1334
Issue: rf#14292
Change-Id: I7b5aad26c14e8508cdbde3fa717434ec0302fab1