[media] vinc: Add control for AE target luminance
Target luminance is one of the major parameter of
autoexposure algorithm. It should be made adjustable to
get best image quality in different lighting conditions.

Previous target luminance was fixed to 110. This value was
chosen by eye for gamma ~ 1. New default value is set to 50.
This corresponds to the relative luminance of 18% gray card
(human vision perceives it as middle gray).

Issue: #IPCAMSW-554
Change-Id: Id3435563dd0a6b12e85e997a044d24119ffe9a75