[media] vinc: Add control for sensor name
To load profile for specific sensor vinc-profile is required
to get sensor name. Currently it is not possible to get sensor
name from vinc driver.

An easy (however not elegant) solution to the problem is to
add a control that allows to get name of sensor connected to
vinc device.

Issue: #IPCAMSW-484
Change-Id: I322e3b9cca7ecfbe763cb9311232bff82132a5e7