regulator: rpi-panel: Serialise operations.
The driver was using the regmap lock to serialise the
individual accesses, but we really need to protect the
timings of enabling the regulators, including any communication
with the Atmel.

Use a mutex within the driver to control overall accesses to
the Atmel, instead of the regmap lock.

Change-Id: I0c07136f93c7111fabd401c73905ffdc4039a5d8
Origin: https://github.com/raspberrypi/linux/commit/6ace7cd753d0
Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>