arm64: dts: elvees: Add snps,dis_u2_susphy_quirk for USB
In a configuration found on ELV-MC03-SMARC + NGFW-CB:

  [MCom-03 USB] --> [SuperSpeed hub] --> [SuperSpeed device],

the device is not detected unless an USB2 device is connected to the
same hub for unknown reason.

Is was found that this issue can be worked around by disabling USB2 PHY
suspend in DWC3 controller.

Issue: #MCOM03BUP-248
Issue: #MCOM03BUP-251
Change-Id: Ic494f3ea3b51439751d7341ae70fae1e6190cf11