mtd: rawnand: arasan: Work around hang on write to DATA_PORT
The MCom-03 version of the Arasan NAND controller was found to hang
Linux on writes to the DATA_PORT register.

It is unknown why adding delay before writes to this register resolves
this issue, but the workaround is good enough for the purposes the
register is used for (e.g. setting features).

Issue: #MCOM03BUP-164
Change-Id: Iaa8abcdb34de67a7cc8ec34e3b5f07dab62f66b5