regulator: rpi-panel: Remove get_brightness hook
The driver was implementing a get_brightness function that
tried to read back the PWM setting of the display to report
as the current brightness.
The controller on the display does not support that, therefore
we end up reporting a brightness of 0, and that confuses
systemd's backlight service.

Remove the hook so that the framework returns the current
brightness automatically.

Change-Id: I1270ff5e93ab7ce008c36e2fe021ad280938a4d6
Origin: https://github.com/raspberrypi/linux/commit/ce646f7446fb
Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>