arm: dts: elvees: mcom03: Fix PCIe ranges
In the Linux 4.19 version of the Designware PCIe driver,
PCIe ranges are configured exclusively in 32-bit mode,
where the range end address is calculated by using
the upper 32 bits from the range
base address (set in Device Tree), and the lower 32 bits
as specified in the range limit address register
(calculated as lower_32_bits(base + size - 1)).

This presents us with two limitations:
1. The PCIe range cannot be larger than 4 GiB, and
2. The PCIe range cannot cross a 4 GiB boundary.

For example, end address for PCIe1 was calculated as follows
before this commit:

  range_base = 0x5_c000_0000 (from DT)
  range_size = 0x1_c000_0000 (from DT)
  range_end = (range_base & GENMASK_ULL(63, 32)) |
              ((range_base + range_size - 1) & GENMASK(31, 0))
            = 0x5_7fff_ffff.

This is lower than the base address, so range mapping
does not work at all.

Change-Id: Ib81dbc5daf1b9f87076ff9e574e4889d838cd7f3
Issue: #MCOM03SW-828