pci: dwc: mcom03: Fixup BARs before initialization
The following error was noticed during PCIe init:

  BAR 0: assigned [mem 0x6c0000000-0x6ffffffff 64bit]
  BAR 0: error updating (high 0x000001 != 0x000000)

This error is caused by the init code for Designware PCIe RC
expecting DBI_RO_WR_EN to be disabled, and writing 0x4 to BAR0.
In actuality, on MCom-03 DBI_RO_WR_EN is enabled on reset, so
the write sets bits [3:0] that are zeroed on reset.
This makes the rest of the PCIe subsystem erroneously believe
that BAR0 is a 64-bit register, and attempt to write
to bits [3:0] of BAR1, which are read-only.

We fix this in two ways:

1. DBI_RO_WR_EN is now disabled before the Designware driver
   writes anything to registers.
2. The sizes of BAR0 and BAR1 are corrected to more sensible
   values - from 1 GiB and 2 GiB that were configured on reset
   to both being 4 KiB. These BARs are not used by Linux, but
   still take up space in PCIe memory ranges.

Issue: #MCOM03SW-1369
Change-Id: Icefe2fade3b3fb55af1f17c17aa76d3b90787d03
Signed-off-by: Vladimir Terekhov <vterekhov@elvees.com>
(cherry picked from commit b62d48833bcc7e4ab6823cff81ad4d93b0a8ed70)