pci_endpoint_test/pci_epf_test: Add checksummed BAR read/write
The BAR tests previously used in pcitest had a fatal flaw: both the
reading and the checking was done on RC. This meant that if the writes
never got to EP (e.g. the place where BAR is mapped has some kind of
RAM backing), the tests would still succeed.

Change-Id: I356f0042cc31bfb84d3753c32be24b73d02a4e61