pinctrl: dwapb: Set proper configs before requesting IRQ
Device drivers usually use devm_request_threaded_irq() to request IRQ,
which assumes that GPIO pin already in 'Software mode' and the
direction is input.

This commit guarantees that GPIO pin/port will be in 'software mode'
and in input mode before locking GPIO pin as IRQ.

Muxing pin to 'Software mode' upon requesting IRQ by
devm_request_threaded_irq() is not handled by the pinctrl driver, this
means that device drivers in DeviceTree should consider that, there is
no collision detection when requesting IRQ. In the other hand
requesting IRQ from the userspace is handled by the pinctrl driver
because gpiolib is requesting the pin before it use it.

Issue: #MCOM03SW-1182
Change-Id: I800969525f347d5d262a70a9b7cc19b6e4917497