pinctrl: dwapb: Fix debugged groups indices
Pinmaping of GPIO differs between GPIO/PINCTRL subsystems if the PINCTRL
device of type 'mux_individual_pins == false'. PINCTRL driver of this
type will have one group with the name gpio{A,B,C,D}_grp.

Requesting a pin from the GPIO subsystem is handled as a request to a
group in the PINCTRL driver, so in order to debug it we need to rely on
port index, not on group selector which represent pin index.

Change-Id: I4ffdec583f9060a49f455ac052953657fecf5081