PCI: dwc: Do not remap invalid res
On MCom-03 and some other platforms when PCIe probes the following
error occurs:

mcom03-pcie 2400000.pcie: invalid resource

This happens because the "atu" is not specified in the Device Tree
and therefore should not be remapped.

This commit is based on upstream commit 6e5ebc96ec65
("PCI: dwc: Do not remap invalid res").

Issue: #MCOM03SW-1369
Change-Id: Ib4fa7e30deb7fd4607d891cc120568448ecd8cf0
Signed-off-by: Vladimir Terekhov <vterekhov@elvees.com>