arm64: dts: elvees: Move I2C4 description into separate dtsi
I2C4 is located in service subsystem and should not be configured
by Linux kernel in case of secure applications are running.
We have to remove it but it is still used in tests and required
for legacy board support. Currently we move it in the separate dtsi
and will remove in future.

Issue: #MCOM03SW-1885
Change-Id: I4aaca7df3458f188fc84202713c7f5c829e22a22
Signed-off-by: Mikhail Konovalov <mkonovalov@elvees.com>
(cherry picked from commit b17022891764941bf77069f74ac88c9fdb2b0086)