arm64: dts: elvees: Move I2C4 description into separate dtsi
I2C4 is located in service subsystem and should not be configured
by Linux kernel in case of secure applications are running.
We have to remove it but it is still used in tests and required
for legacy board support. Currently we move it in the separate dtsi
and will remove in future.

Issue: #MCOM03SW-1885
Change-Id: I55bee08cdc45cf9fef50ca1839d208d334fe7c92