arm64: dts: elvees: Add spi0 fifo-size property
The MCom-03 SoC SPI controller has a hardware bug with overrun error
occurs on RX FIFO. This commit is workaround of this bug.

The MCom-03 SoC includes a DW apb_ssi v4.01* controller which is
documented to have a 64 words deep TX and RX FIFO. The FIFO
length detection in spi_hw_init() correctly detects this value.
However, when the controller RX FIFO is filled up to 64 entries
(RXFLR = 64), an RX FIFO overrun error occurs. This likely due to a
hardware bug which can be avoided by force setting the fifo_len
field of struct dw_spi to 63.

Issue: #MCOM03SW-1636
Issue: #MCOM03-1879
Change-Id: I243546356144fcca64dfeaefe45eebd14e2c84ba