arm64: dts: elvees: rockpi-n10: Add aliases for cam0/1 and lcd
After adding support for i2c4, the i2c port number for the camera
is changed. Adding aliases will allow to fix the port numbers for
cameras and lcd. The port number 6 for camera 0 corresponds to the
number set in the configuration file for the board in the felix repo.

Issue: #MCOM03SW-2347
Change-Id: I04f029a65d3d42abb97224342d292a9e94603819
Signed-off-by: Julia Reshetnikova <ureshetnikova@elvees.com>