net: mfbsp-can: Add support for controlling PHY ICs
Controlling CAN PHY ICs is required when it's Standby and Enable pins
are connected to GPIO, and Linux has to control it, how it's done, e.g,
on MCom03-BuB with ATA6561 CAN PHY transceiver. Thus Linux must turn on
PHY ICs when working with CAN and vice-versa.

Issue: #MCOM03SW-2337
Change-Id: I002e5db2233e1af603661bfbbc7d43029ae4962b