arm64: dts: elvees: ecam03: Add DTS common for all ECAM03* r2.0
Currently at first boot U-Boot reads 'compatible' from U-Boot DTB and
tries to find the corresponding Linux DTB. For ECAM03* r2.0, the problem
is that U-Boot uses a common DTB for all ECAM03* r2.0, which has no
analogue in Linux. As a result at first boot it is required to manually
override the 'board' variable in U-Boot terminal, but this is inconvenient
as it requires UART connection.

This commit adds common for all ECAM03* r2.0 boards Linux DTB to get rid
of the problem.

Issue: #MCOM03SW-2382
Change-Id: Ief165c9a8eccbe58e678f3a94ac6609f04a473bc
Signed-off-by: Dmitriy Zagrebin <dzagrebin@elvees.com>