arm64: dts: elvees: Fix elcores compatible list order
Accodring to the Device Tree docs the first string of compatible
list should specify the exact device that the node represents in
the form "<manufacturer>,<model>". The following strings represent
other devices that the device is compatible with.

So "simple-bus" compatible should be in the last position.

Issue: #MCOM03SW-2981
Change-Id: Iccf6fc000369e3d0511c34129c922d2ff694bca4
Signed-off-by: Georgy Macharadze <gmacharadze@elvees.com>