[media] vinc: Correct controls for non-bayer sensors
Sensors that send data in RGB and YCbCr formats
(e.g. parallel OV2640) have white auto-balance,
auto-exposure and auto-gain controls.
The driver enables VINC auto-controls by default.
For non-bayer sensors, this leads to artifacts.
The commit adds a check for the presence of auto controls
in the sensor, and if available, turns them on,
and turns off the VIN auto control.
If the data format is from sensor YCbCr, auto controls in the VINС
is prohibited.

Issue: #MCOM02SW-2256
Change-Id: I9cd9bc1cd64417b5789fa8603936e86cedc55c1b