elvees-swic: Fix passing NULL pointer
The dma_alloc_coherent()/dma_alloc_attrs() functions do not check the
NULL pointer of pdata->dev, which is passed as the first argument.
In the current kernel version this situation does not lead to the
observed error, but we cannot guarantee correct behavior of driver.

To fix this problem it need to set the pdata->dev = &pdev->dev before
first use.

The problem was found when migrating the driver from 4.4 to 5.4.

Issue: #MCOM02SW-2163
Change-Id: I4416b528ada14543593806e8fc8be3670277d301