arm64: dts: elvees: Decrease QSPIx speed to 50 MHz
Some chips have a trouble with access to QSPI flash on 66 MHz and
higher. QSPI1 works with 3.3 V voltages.

At the 66 MHz clock frequency we see, that a voltage doesn't reach
the zero for QSPI1.

QSPI0 works with 1.8 V voltages and we have the issue with it quite
rarely, but we should decrease clock for it too to increase robustness.

Issue: #MCOM03SW-3753
Change-Id: I5680a0947c4c27389fcbdbca98c7ec76b0bdd156