drm/bridge: lt9611uxc: Add workaround for EDID interrupt loss
ELV-MC03-SMARC boards use an GPIO port expander with interrupt output
with limited functionality (lack of latch register) - XL9555. It does
not store information about interrupts that occur on the input GPIO
ports.

The interrupt from the LT9611UXC is connected to the input GPIO ports
of XL9555 and in some cases may be lost during interrupt handling.

This commit modifies detect callback of drm_bridge control functions,
which checks if anything is attached to the bridge output. In this
callback, information about the HDMI cable connection is read from
a register that also contains EDID ready status. This commit get
this status and update internal driver status.

Issue: #MCOM03SW-3431
Change-Id: I6c54fcc4dc52937df4d34aa30ca5a7971d0aa8e1
Signed-off-by: Oleg Ivanov <oaivanov@elvees.com>